+++ src/lib/CLI.ts
@@ -8,17 +8,18 @@ import {
 } from "fs";
 import { Agent } from "https";
 import { homedir } from "os";
-import { load as yamlLoad } from "js-yaml";
+import { load as yamlLoad, dump as yamlDump } from "js-yaml";
 import { Automatiqal } from "automatiqal";
 import { varLoader } from "@informatiqal/variables-loader";
 import { printTable, Table } from "console-table-printer";
 
-import { IRunBook } from "automatiqal/dist/RunBook/RunBook.interfaces";
+import { IRunBook, ITask } from "automatiqal/dist/RunBook/RunBook.interfaces";
 import { ITaskResult } from "automatiqal/dist/RunBook/Runner";
 import { Logger } from "./Logger";
 
 import { IArguments } from "./interfaces";
 import * as path from "path";
+import { schemaURL } from "./sample";
 
 export class AutomatiqalCLI {
   private argv: IArguments;
@@ -29,7 +30,8 @@ export class AutomatiqalCLI {
   private automatiqal: Automatiqal;
   private rawRunBook: string;
   private runbookVariablesList: string[];
-  private variablesValues: { [x: string]: string };
+  // private runbookVariablesValues: { [k: string]: any } = {};
+  private variablesValues: { [x: string]: string | number | boolean };
   private emitterMessages: string[];
   private specialVariables = [
     "GUID",
@@ -44,7 +46,7 @@ export class AutomatiqalCLI {
     this.argv = argv;
     this.result = [];
     this.rawRunBook = downloadedRunbook;
-    this.logger = Logger.getInstance();
+    this.logger = Logger.getInstance(argv.summary || argv.s);
 
     // if the runbook is downloaded then clean it first
     // by removing commented and empty lines
@@ -57,6 +59,31 @@ export class AutomatiqalCLI {
         this.rawRunBook = this.readRunbookFile();
       } catch (e) {
         this.logger.error(e.message, 1000);
+      }
+    }
+
+    this.runbookSet();
+    this.importExternalYamlFiles();
+    this.rawRunBook = JSON.stringify(this.runBook);
+
+    if (this.argv.compile) {
+      const yamlRunbook = yamlDump(this.runBook);
+      const output = `${schemaURL}\n\n${yamlRunbook}`;
+
+      if (this.argv.compile == true) {
+        console.log(output);
+        process.exit(1);
+      }
+
+      if (typeof this.argv.compile == "string") {
+        try {
+          writeFileSync(this.argv.compile, output);
+          console.log(`Compiled schema was saved in "${this.argv.compile}"`);
+          process.exit(1);
+        } catch (e) {
+          console.log(e.message);
+          process.exit(1);
+        }
       }
     }
 
@@ -129,10 +156,15 @@ export class AutomatiqalCLI {
       this.argv.env ||
       this.argv.i ||
       this.argv.inline
-    )
-      this.replaceVariables();
+    ) {
+      this.rawRunBook = this.replaceVariables(
+        this.rawRunBook,
+        this.variablesValues
+      );
+      this.runBook = JSON.parse(this.rawRunBook);
+    }
 
-    this.runbookSet();
+    this.checkExportPaths();
 
     if (argv.listvars || argv.l) {
       this.printDefinedVariables();
@@ -151,7 +183,13 @@ export class AutomatiqalCLI {
     if (!this.argv.c && !this.argv.connect) this.readBuffers();
 
     try {
-      this.automatiqal = new Automatiqal(this.runBook, this.httpsAgent);
+      const disableSchemaValidation =
+        this.argv.disableValidation || this.argv.d;
+
+      this.automatiqal = new Automatiqal(this.runBook, {
+        httpsAgent: this.httpsAgent,
+        disableSchemaValidation: disableSchemaValidation || false,
+      });
     } catch (e) {
       if (e.context) throw new Error(e.context);
       if (e.message) throw new Error(e.message);
@@ -189,8 +227,6 @@ export class AutomatiqalCLI {
         this.logger.error(e.message, 1003);
       }
     }
-
-    this.checkExportPaths(this.runBook.tasks);
 
     if (this.argv.c || this.argv.connect) {
       this.runBook.tasks = [
@@ -277,8 +313,9 @@ export class AutomatiqalCLI {
         _this.logger.taskEntry(
           b.timings.start,
           b.timings.end,
-          `${b.timings.totalSeconds}(s)`,
-          b.task.name,
+          `${b.timings.totalSeconds}`,
+          b.task.name.replace(/(.{27})..+/, "$1..."),
+          `${(b.data as []).length}`,
           b.status
         );
       }
@@ -289,6 +326,7 @@ export class AutomatiqalCLI {
     this.automatiqal.emitter.on("runbook:log", function (l) {});
 
     this.automatiqal.emitter.on("error", function (errorMessage) {
+      //TODO: here. exit?
       _this.logger.error(errorMessage);
     });
   }
@@ -387,17 +425,23 @@ export class AutomatiqalCLI {
   /**
    * @description replace all runbook variables with their content
    */
-  private replaceVariables() {
-    Object.entries(this.variablesValues).forEach(([varName, varValue]) => {
+  private replaceVariables(
+    text: string,
+    vars: { [x: string]: string | number | boolean }
+  ) {
+    Object.entries(vars).forEach(([varName, varValue]) => {
       try {
         const v = "\\$\\{" + varName + "\\}";
         const re = new RegExp(v, "g");
 
-        this.rawRunBook = this.rawRunBook.replace(re, varValue.toString());
+        // this.runbookVariablesValues[varName] = varValue;
+        text = text.replace(re, varValue.toString());
       } catch (e) {
         this.logger.error(e.message, 9999);
       }
     });
+
+    return text;
   }
 
   // TODO: dont like this much. Refactor at some point
@@ -451,22 +495,23 @@ export class AutomatiqalCLI {
     this.logger.taskEntry(
       "START TIME".padEnd(24, " "),
       "END TIME".padEnd(24, " "),
-      "DURATION",
+      "DURATION(s)",
       "TASK NAME".padEnd(30, " "),
+      "ENTITIES".padEnd(10, " "),
       "STATUS"
     );
   }
 
-  private checkExportPaths(tasks: any) {
+  private checkExportPaths() {
     let nonExistingPaths = [];
 
-    tasks.forEach((task) => {
-      if (task.details?.location) {
-        const isValidPath = existsSync(task.details?.location);
+    this.runBook.tasks.forEach((task) => {
+      if (task.details?.["location"]) {
+        const isValidPath = existsSync(task.details?.["location"]);
 
         if (!isValidPath)
           nonExistingPaths.push(
-            `Initial check: Export path for task "${task.name}" do not exists "${task.details.location}"`
+            `Initial check: Export path for task "${task.name}" do not exists "${task.details["location"]}"`
           );
       }
     });
@@ -542,5 +587,52 @@ export class AutomatiqalCLI {
     table.printTable();
 
     console.log("");
+  }
+
+  private importExternalYamlFiles(): void {
+    let parsed = false;
+
+    while (parsed == false) {
+      let haveImport = false;
+
+      for (let i = 0; i < this.runBook.tasks.length; i++) {
+        if (this.runBook.tasks[i]["import"]) {
+          const externalFileContent = this.readImportFile(
+            this.runBook.tasks[i]["import"]["path"] ||
+              this.runBook.tasks[i]["import"],
+            this.runBook.tasks[i]["import"]["vars"] || []
+          );
+
+          this.runBook.tasks.splice(i, 1, ...externalFileContent);
+
+          haveImport = true;
+        }
+      }
+
+      if (haveImport == false) parsed = true;
+    }
+  }
+
+  private readImportFile(
+    path: string,
+    variables: { [k: string]: string | number | boolean }[]
+  ) {
+    const exists = existsSync(path);
+
+    const v = {};
+    variables.map((v1) => {
+      Object.entries(v1).forEach(([key, value]) => {
+        v[key] = value;
+      });
+    });
+
+    if (!exists) this.logger.error(`Import file not found: ${path}`);
+
+    let fileContent = readFileSync(path).toString();
+    fileContent = this.replaceVariables(fileContent, v);
+
+    const parsedContent = yamlLoad(fileContent);
+
+    return parsedContent as ITask[];
   }
 }